set SUPERBLOCK_STORING_WINDOW to 2 hours
we wan't protection against a bad agent from sending invalid superblocks to the contracts before agents get a chance to challenge. So we want locally stored superblocks stored up to 2 hours in advance of being sent to the contract so that when new SB are added it can be checked against locally.

Now that we have stored them 2 hours in advance that means we should have the SB stored locally way in advance from when the contract shoudl get a new SB. This means if an attacker submits a bad superblock ahead of schedule (2 hour descrepency so its obvious without doubt) then challengers shoudl see that the contract SB and the local SB doesn't match and challenge it.